refactor(api): split agent route monolith into sub-package
Split observal-server/api/routes/agent.py (1707 lines) into focused modules:

- agent/helpers.py: _load_agent, _agent_to_response, _resolve_component_names
- agent/crud.py: create, list, get, update, delete, archive/unarchive
- agent/install.py: install, download_stats, traces, resolve, manifest, validate
- agent/draft.py: save_draft, update_draft, start/cancel edit, submit_draft

Uses shared _router.py pattern so main.py requires no import changes.
Test mock.patch targets updated to reference new sub-module paths.
