+++ observal-server/tests/test_agent_pull.py
@@ -153,7 +153,7 @@ def test_agent_response_schema_has_latest_version_field():
 
 def test_agent_to_response_populates_latest_approved_version():
     """_agent_to_response must set latest_approved_version from agent.versions."""
-    from api.routes.agent import _agent_to_response
+    from api.routes.agent.helpers import _agent_to_response
 
     agent = _make_agent(with_approved_version=True)
     # Ensure the approved version is in agent.versions list
@@ -172,7 +172,7 @@ def test_agent_to_response_populates_latest_approved_version():
 
 def test_agent_to_response_latest_approved_version_none_when_no_approved():
     """_agent_to_response sets latest_approved_version=None when no approved version."""
-    from api.routes.agent import _agent_to_response
+    from api.routes.agent.helpers import _agent_to_response
 
     agent = _make_agent(with_approved_version=False)
     # Add a pending version (not approved)
@@ -204,7 +204,7 @@ def test_agent_to_response_latest_approved_version_none_when_no_approved():
 
 def test_agent_to_response_populates_latest_version_string():
     """_agent_to_response sets latest_version to agent.version string."""
-    from api.routes.agent import _agent_to_response
+    from api.routes.agent.helpers import _agent_to_response
 
     agent = _make_agent(with_approved_version=True)
 
@@ -227,7 +227,7 @@ async def test_install_agent_no_latest_version_returns_400():
     """install_agent returns 400 when agent has no approved/published version."""
     from fastapi import HTTPException
 
-    from api.routes.agent import install_agent
+    from api.routes.agent.install import install_agent
     from schemas.agent import AgentInstallRequest
 
     agent = _make_agent(with_approved_version=False)
@@ -245,9 +245,9 @@ async def test_install_agent_no_latest_version_returns_400():
     db.commit = AsyncMock()
 
     with (
-        patch("api.routes.agent._load_agent", new=AsyncMock(return_value=agent)),
-        patch("api.routes.agent.get_effective_agent_permission", return_value="owner"),
-        patch("api.routes.agent.audit", new=AsyncMock()),
+        patch("api.routes.agent.install._load_agent", new=AsyncMock(return_value=agent)),
+        patch("api.routes.agent.install.get_effective_agent_permission", return_value="owner"),
+        patch("api.routes.agent.install.audit", new=AsyncMock()),
         patch("services.download_tracker.record_agent_download", new=AsyncMock()),
         pytest.raises(HTTPException) as exc,
     ):
@@ -266,7 +266,7 @@ async def test_install_agent_no_latest_version_returns_400():
 @pytest.mark.asyncio
 async def test_install_agent_with_approved_version_succeeds():
     """install_agent returns 200 with config_snippet when agent has an approved version."""
-    from api.routes.agent import install_agent
+    from api.routes.agent.install import install_agent
     from schemas.agent import AgentInstallRequest
 
     user = _make_user()
@@ -288,11 +288,11 @@ async def test_install_agent_with_approved_version_succeeds():
     fake_config = {"mcpServers": {}, "rules": "# ruffchecker\n"}
 
     with (
-        patch("api.routes.agent._load_agent", new=AsyncMock(return_value=agent)),
-        patch("api.routes.agent.get_effective_agent_permission", return_value="owner"),
-        patch("api.routes.agent.generate_agent_config", return_value=fake_config),
-        patch("api.routes.agent.emit_registry_event"),
-        patch("api.routes.agent.audit", new=AsyncMock()),
+        patch("api.routes.agent.install._load_agent", new=AsyncMock(return_value=agent)),
+        patch("api.routes.agent.install.get_effective_agent_permission", return_value="owner"),
+        patch("api.routes.agent.install.generate_agent_config", return_value=fake_config),
+        patch("api.routes.agent.install.emit_registry_event"),
+        patch("api.routes.agent.install.audit", new=AsyncMock()),
         patch("api.routes.config.derive_endpoints", return_value={"api": "http://localhost:8000", "otlp_http": ""}),
         patch("services.download_tracker.record_agent_download", new=AsyncMock()),
     ):
