+++ tests/test_draft_workflow.py
@@ -161,7 +161,7 @@ class TestDraftSave:
 
     @pytest.mark.asyncio
     @patch("services.agent_snapshot.build_yaml_snapshot", new=AsyncMock(return_value="snapshot"))
-    @patch("api.routes.agent._load_agent")
+    @patch("api.routes.agent.draft._load_agent")
     async def test_creates_agent_with_draft_status(self, mock_load):
         """POST /agents/draft creates an agent in draft status."""
         user = _user()
@@ -182,7 +182,7 @@ class TestDraftSave:
 
     @pytest.mark.asyncio
     @patch("services.agent_snapshot.build_yaml_snapshot", new=AsyncMock(return_value="snapshot"))
-    @patch("api.routes.agent._load_agent")
+    @patch("api.routes.agent.draft._load_agent")
     async def test_response_includes_agent_fields(self, mock_load):
         """Draft response includes id, name, and status fields."""
         user = _user()
@@ -212,7 +212,7 @@ class TestDraftUpdate:
 
     @pytest.mark.asyncio
     @patch("services.agent_snapshot.build_yaml_snapshot", new=AsyncMock(return_value="snapshot"))
-    @patch("api.routes.agent._load_agent")
+    @patch("api.routes.agent.draft._load_agent")
     async def test_updates_draft_fields(self, mock_load):
         """PUT /agents/{id}/draft updates the draft agent."""
         user = _user()
@@ -230,7 +230,7 @@ class TestDraftUpdate:
         db.commit.assert_awaited()
 
     @pytest.mark.asyncio
-    @patch("api.routes.agent._load_agent")
+    @patch("api.routes.agent.draft._load_agent")
     async def test_rejects_update_on_non_draft(self, mock_load):
         """Updating an approved agent returns 400."""
         user = _user()
@@ -247,7 +247,7 @@ class TestDraftUpdate:
         assert r.status_code == 400
 
     @pytest.mark.asyncio
-    @patch("api.routes.agent._load_agent")
+    @patch("api.routes.agent.draft._load_agent")
     async def test_rejects_update_by_non_owner(self, mock_load):
         """Non-owner cannot update a draft agent (403)."""
         user = _user()
@@ -274,9 +274,9 @@ class TestDraftSubmit:
 
     @pytest.mark.asyncio
     @patch("services.agent_snapshot.build_yaml_snapshot", new=AsyncMock(return_value="snapshot"))
-    @patch("api.routes.agent.emit_registry_event")
-    @patch("api.routes.agent._resolve_component_names")
-    @patch("api.routes.agent._load_agent")
+    @patch("api.routes.agent.draft.emit_registry_event")
+    @patch("api.routes.agent.draft._resolve_component_names")
+    @patch("api.routes.agent.draft._load_agent")
     async def test_transitions_to_pending(self, mock_load, mock_resolve, mock_emit):
         """POST /agents/{id}/submit transitions a draft to pending status."""
         user = _user()
@@ -294,9 +294,9 @@ class TestDraftSubmit:
 
     @pytest.mark.asyncio
     @patch("services.agent_snapshot.build_yaml_snapshot", new=AsyncMock(return_value="snapshot"))
-    @patch("api.routes.agent.emit_registry_event")
-    @patch("api.routes.agent._resolve_component_names")
-    @patch("api.routes.agent._load_agent")
+    @patch("api.routes.agent.draft.emit_registry_event")
+    @patch("api.routes.agent.draft._resolve_component_names")
+    @patch("api.routes.agent.draft._load_agent")
     async def test_submit_response_includes_status(self, mock_load, mock_resolve, mock_emit):
         """Submit response body includes the new pending status."""
         user = _user()
@@ -322,7 +322,7 @@ class TestDraftSubmitNotDraft:
     """Test submitting a non-draft agent returns an error."""
 
     @pytest.mark.asyncio
-    @patch("api.routes.agent._load_agent")
+    @patch("api.routes.agent.draft._load_agent")
     async def test_submit_pending_returns_400(self, mock_load):
         """Submitting a pending agent returns 400."""
         user = _user()
@@ -336,7 +336,7 @@ class TestDraftSubmitNotDraft:
         assert r.status_code == 400
 
     @pytest.mark.asyncio
-    @patch("api.routes.agent._load_agent")
+    @patch("api.routes.agent.draft._load_agent")
     async def test_submit_active_returns_400(self, mock_load):
         """Submitting an active agent returns 400."""
         user = _user()
@@ -350,7 +350,7 @@ class TestDraftSubmitNotDraft:
         assert r.status_code == 400
 
     @pytest.mark.asyncio
-    @patch("api.routes.agent._load_agent")
+    @patch("api.routes.agent.draft._load_agent")
     async def test_submit_not_found_returns_404(self, mock_load):
         """Submitting a nonexistent agent returns 404."""
         user = _user()
